dyninst/decode: move hmap logic to a file
